--- routes/authRoutes.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { loginUser } = require('../controllers/authController');
-
-// POST /user/login
-router.post('/login', loginUser);
-
-module.exports = router;
